requirements.txt
torch
torchvision
opencv-python
mediapipe
face-recognition
PyYAML
numpy
pygame
Flask
face_recognition